backend/requirements.txt
alembic==1.14.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
openpyxl==3.1.5
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
